+++ CHANGELOG.md
@@ -24,6 +24,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - HCI Dump Dispatch: allow to use multiple HCI Dump implementations
 - HCI Dump: support ISO packets in PacketLogger format
 - Immediate Alert Service Server and Client: added
+- Link Loss Service Server and Client: added
 - btstack_util: safe wrappers for snprintf
 
 ### Fixed